refactor: Move types to separate crate (#179)
