+++ benchmarks/src/main.rs
@@ -1,7 +1,8 @@
 use bitcoin::consensus::Decodable;
 use bitcoin::Block as BitcoinBlock;
-use ic_btc_canister::{types::Block, with_state_mut};
+use ic_btc_canister::with_state_mut;
 use ic_btc_interface::{Config, Network};
+use ic_btc_types::Block;
 use ic_cdk_macros::{init, query};
 use std::cell::RefCell;
