+++ canister/src/api/metrics.rs
@@ -1,4 +1,5 @@
-use crate::{metrics::InstructionHistogram, state, types::HttpResponse, with_state};
+use crate::{metrics::InstructionHistogram, state, with_state};
+use ic_btc_types::HttpResponse;
 use ic_cdk::api::time;
 use ic_metrics_encoder::MetricsEncoder;
 use serde_bytes::ByteBuf;
