+++ canister/src/block_header_store.rs
@@ -1,7 +1,6 @@
-use crate::{
-    memory::Memory,
-    types::{Block, BlockHash, BlockHeaderBlob},
-};
+use crate::memory::Memory;
+use ic_btc_types::{Block, BlockHash, BlockHeaderBlob};
+
 use bitcoin::consensus::{Decodable, Encodable};
 use bitcoin::BlockHeader;
 use ic_btc_interface::Height;
