+++ canister/src/test_utils.rs
@@ -1,7 +1,4 @@
-use crate::{
-    genesis_block,
-    types::{into_bitcoin_network, Address, Block, OutPoint, Transaction},
-};
+use crate::genesis_block;
 use bitcoin::{
     hashes::Hash, secp256k1::rand::rngs::OsRng, secp256k1::Secp256k1, Address as BitcoinAddress,
     BlockHeader, PublicKey, Script, WScriptHash,
@@ -10,6 +7,7 @@ use ic_btc_interface::Network;
 use ic_btc_test_utils::{
     BlockBuilder as ExternalBlockBuilder, TransactionBuilder as ExternalTransactionBuilder,
 };
+use ic_btc_types::{into_bitcoin_network, Address, Block, OutPoint, Transaction};
 use ic_stable_structures::{BoundedStorable, Memory, StableBTreeMap};
 use proptest::prelude::RngCore;
 use std::str::FromStr;
@@ -139,7 +137,7 @@ pub fn is_stable_btreemap_equal<
     true
 }
 
-/// A wrapper around `ic_btc_test_utils::BlockBuilder` that returns `crate::types::Block`
+/// A wrapper around `ic_btc_test_utils::BlockBuilder` that returns `ic_btc_types::Block`
 /// as opposed to `bitcoin::Block`.
 pub struct BlockBuilder {
     builder: ExternalBlockBuilder,
@@ -176,7 +174,7 @@ impl BlockBuilder {
 }
 
 /// A wrapper around `ic_btc_test_utils::TransactionBuilder` that returns
-/// `crate::types::Transaction` as opposed to `bitcoin::Transaction`.
+/// `ic_btc_types::Transaction` as opposed to `bitcoin::Transaction`.
 pub struct TransactionBuilder {
     builder: ExternalTransactionBuilder,
 }
