+++ types/src/lib.rs
@@ -1,4 +1,3 @@
-use crate::state::OUTPOINT_SIZE;
 use bitcoin::{
     util::uint::Uint256, Address as BitcoinAddress, Block as BitcoinBlock,
     BlockHeader as BitcoinBlockHeader, Network as BitcoinNetwork, OutPoint as BitcoinOutPoint,
@@ -22,6 +21,9 @@ use std::{
     str::FromStr,
 };
 
+mod constants;
+pub use crate::constants::{max_target, OUTPOINT_SIZE};
+
 // The expected length in bytes of the page.
 const EXPECTED_PAGE_LENGTH: usize = 72;
 
@@ -31,7 +33,7 @@ pub struct Block {
     block_header: BlockHeader,
     transactions: Vec<Transaction>,
 
-    #[cfg(test)]
+    #[cfg(feature = "test-utils")]
     pub mock_difficulty: Option<u64>,
 }
 
@@ -44,7 +46,7 @@ impl Block {
                 .map(|tx| Transaction::new(tx.clone()))
                 .collect(),
             block_header: BlockHeader::new(block.header),
-            #[cfg(test)]
+            #[cfg(feature = "test-utils")]
             mock_difficulty: None,
         }
     }
@@ -62,18 +64,19 @@ impl Block {
     }
 
     pub fn difficulty(&self, network: Network) -> u64 {
-        #[cfg(test)]
+        #[cfg(feature = "test-utils")]
         if let Some(difficulty) = self.mock_difficulty {
             return difficulty;
         }
 
         Self::target_difficulty(network, self.header().target())
     }
-    #[cfg(test)]
+
+    #[cfg(feature = "test-utils")]
     pub fn consensus_encode(&self, buffer: &mut Vec<u8>) -> Result<usize, std::io::Error> {
         use bitcoin::consensus::Encodable;
         let bitcoin_block = BitcoinBlock {
-            header: self.header().clone(),
+            header: *self.header(),
             txdata: self
                 .transactions
                 .clone()
@@ -88,7 +91,7 @@ impl Block {
     // The definition here corresponds to what is referred as "bdiff" in
     // https://en.bitcoin.it/wiki/Difficulty
     fn target_difficulty(network: Network, target: Uint256) -> u64 {
-        (ic_btc_validation::max_target(&into_bitcoin_network(network)) / target).low_u64()
+        (crate::max_target(&into_bitcoin_network(network)) / target).low_u64()
     }
 }
 
@@ -186,7 +189,7 @@ impl PartialEq for Transaction {
     }
 }
 
-#[cfg(test)]
+#[cfg(feature = "test-utils")]
 impl From<Transaction> for bitcoin::Transaction {
     fn from(tx: Transaction) -> Self {
         tx.tx
@@ -215,7 +218,7 @@ impl From<&BitcoinOutPoint> for OutPoint {
     }
 }
 
-#[cfg(test)]
+#[cfg(feature = "test-utils")]
 impl From<OutPoint> for bitcoin::OutPoint {
     fn from(outpoint: OutPoint) -> Self {
         use bitcoin::hashes::Hash;
@@ -1037,9 +1040,9 @@ fn target_difficulty() {
     assert_eq!(
         Block::target_difficulty(
             Network::Regtest,
-            crate::test_utils::BlockBuilder::genesis()
+            ic_btc_test_utils::BlockBuilder::genesis()
                 .build()
-                .header()
+                .header
                 .target()
         ),
         1
