+++ validation/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 bitcoin = "0.28.1"
+ic-btc-types = { path = "../types" }
 
 [dev-dependencies]
 csv = "1.1"
