+++ validation/src/lib.rs
@@ -1,7 +1,7 @@
 mod constants;
 mod header;
 
-pub use crate::constants::max_target;
 pub use crate::header::{validate_header, HeaderStore, ValidateHeaderError};
+pub use ic_btc_types;
 
 type BlockHeight = u32;
